Enhancement: Re-organize st2tests pack fixtures(#2301)
