+++ conf/st2.tests.conf
@@ -52,7 +52,7 @@ collection_interval = 1
 
 [content]
 system_packs_base_path =
-packs_base_paths = st2tests/st2tests/fixtures/
+packs_base_paths = st2tests/st2tests/fixtures/packs/
 
 [syslog]
 host = 127.0.0.1
