+++ st2api/tests/unit/controllers/v1/test_executions_filters.py
@@ -22,7 +22,7 @@ import six
 from six.moves import http_client
 
 from tests import FunctionalTest
-from st2tests.fixtures import executions as fixture
+from st2tests.fixtures.packs import executions as fixture
 from st2tests.fixtures import history_views
 from st2common.util import isotime
 from st2common.util import date as date_utils
