+++ st2common/tests/unit/test_triggers_registrar.py
@@ -19,7 +19,7 @@ import st2common.bootstrap.triggersregistrar as triggers_registrar
 from st2common.persistence.trigger import Trigger
 from st2common.persistence.trigger import TriggerType
 from st2tests.base import CleanDbTestCase
-from st2tests.fixturesloader import get_fixtures_base_path
+from st2tests.fixturesloader import get_fixtures_packs_base_path
 
 __all__ = [
     'TriggersRegistrarTestCase'
@@ -31,18 +31,18 @@ class TriggersRegistrarTestCase(CleanDbTestCase):
         trigger_type_dbs = TriggerType.get_all()
         self.assertEqual(len(trigger_type_dbs), 0)
 
-        packs_base_path = get_fixtures_base_path()
+        packs_base_path = get_fixtures_packs_base_path()
         count = triggers_registrar.register_triggers(packs_base_paths=[packs_base_path])
-        self.assertEqual(count, 3)
+        self.assertEqual(count, 2)
 
         # Verify TriggerTypeDB and corresponding TriggerDB objects have been created
         trigger_type_dbs = TriggerType.get_all()
         trigger_dbs = Trigger.get_all()
-        self.assertEqual(len(trigger_type_dbs), 3)
-        self.assertEqual(len(trigger_dbs), 3)
+        self.assertEqual(len(trigger_type_dbs), 2)
+        self.assertEqual(len(trigger_dbs), 2)
 
     def test_register_triggers_from_pack(self):
-        base_path = get_fixtures_base_path()
+        base_path = get_fixtures_packs_base_path()
         pack_dir = os.path.join(base_path, 'dummy_pack_1')
 
         trigger_type_dbs = TriggerType.get_all()
