+++ st2tests/st2tests/fixtures/packs/dummy_pack_1/actions/my_action.yaml
