+++ st2tests/st2tests/fixtures/packs/dummy_pack_1/aliases/alias2.yaml
