+++ st2tests/st2tests/fixtures/packs/dummy_pack_1/policies/policy_3.yaml
