+++ st2tests/st2tests/fixtures/packs/dummy_pack_1/triggers/git_commit.yaml
