+++ st2tests/st2tests/fixtures/packs/dummy_pack_2/actions/my_action.py
