+++ st2tests/st2tests/fixtures/packs/dummy_pack_3/actions/my_action.yaml
