+++ st2tests/st2tests/fixtures/packs/dummy_pack_3/pack.yaml
