+++ st2tests/st2tests/fixtures/packs/dummy_pack_4/config.schema.yaml
