+++ st2tests/st2tests/fixtures/packs/dummy_pack_5/config.yaml
