+++ st2tests/st2tests/fixtures/packs/executions/trigger_type.yaml
