+++ st2tests/st2tests/fixtures/packs/pack_invalid_requirements/pack.yaml
