+++ st2tests/st2tests/fixturesloader.py
@@ -132,6 +132,10 @@ def get_fixtures_base_path():
     return os.path.join(os.path.dirname(__file__), 'fixtures')
 
 
+def get_fixtures_packs_base_path():
+    return os.path.join(os.path.dirname(__file__), 'fixtures/packs')
+
+
 def get_resources_base_path():
     return os.path.join(os.path.dirname(__file__), 'resources')
